+++ views/addColumnMigration.php
@@ -7,21 +7,19 @@
 /* @var $table string the name table */
 /* @var $fields array the fields */
 
+preg_match('/^add_(.+)_to_(.+)$/', $name, $matches);
+$columns = $matches[1];
+
 echo "<?php\n";
 ?>
 
 use yii\db\Migration;
 
 /**
- * Handles adding the columns <?php
-foreach ($fields as $field):
-    if ($field == end($fields)):
-        echo "`{$field['property']}`\n";
-    else:
-        echo "`{$field['property']}`, ";
-    endif;
-endforeach;?>
- * for table `<?= $table ?>`.
+ * Handles adding <?= $columns ?> to table `<?= $table ?>`.
+<?= $this->render('_foreignTables', [
+     'foreignKeys' => $foreignKeys
+ ]) ?>
  */
 class <?= $className ?> extends Migration
 {
@@ -30,9 +28,12 @@ class <?= $className ?> extends Migration
      */
     public function up()
     {
-<?php foreach ($fields as $field): ?>
-        $this->addColumn(<?= "'$table', '" . $field['property'] . "', \$this->" . $field['decorators'] ?>);
-<?php endforeach; ?>
+<?= $this->render('_addColumns', [
+    'table' => $table,
+    'fields' => $fields,
+    'foreignKeys' => $foreignKeys,
+])
+?>
     }
 
     /**
@@ -40,8 +41,11 @@ class <?= $className ?> extends Migration
      */
     public function down()
     {
-<?php foreach ($fields as $field): ?>
-        $this->dropColumn(<?= "'$table', '" . $field['property'] . "'" ?>);
-<?php endforeach; ?>
+<?= $this->render('_dropColumns', [
+    'table' => $table,
+    'fields' => $fields,
+    'foreignKeys' => $foreignKeys,
+])
+?>
     }
 }